Improved separation of concerns in UI code
