+++ lib/ipc/main.js
@@ -46,7 +46,7 @@ ipcMain.on('add-to-playlist', function ipcAddToPlaylist(context, data) {
   const player = WindowManager.instances.player.webContents;
   const add = WindowManager.instances.add;
 
-  player.executeJavaScript('RS.Player.populatePlaylist()', true);
+  player.executeJavaScript('RS.Playback.populatePlaylist()', true);
   player.executeJavaScript('RS.displayNotification("' + notification + '");', true);
 
   add.close();
