+++ lib/windowManager.js
@@ -114,9 +114,9 @@ const definitions = {
     template: 'add',
     opts: {
       width: 400,
-      height: 360,
-      minWidth: 320,
-      minHeight: 300,
+      height: 300,
+      minWidth: 340,
+      minHeight: 280,
       show: false,
       showOnReady: true,
       frame: true,
@@ -128,9 +128,9 @@ const definitions = {
     template: 'settings',
     opts: {
       width: 400,
-      height: 360,
-      minWidth: 320,
-      minHeight: 300,
+      height: 300,
+      minWidth: 340,
+      minHeight: 280,
       show: false,
       showOnReady: true,
       frame: true,
@@ -140,10 +140,10 @@ const definitions = {
   help: {
     template: 'help',
     opts: {
-      width: 500,
-      height: 600,
-      minWidth: 320,
-      minHeight: 300,
+      width: 560,
+      height: 640,
+      minWidth: 450,
+      minHeight: 380,
       show: false,
       showOnReady: true,
       frame: true,
