+++ views/add.html
@@ -7,33 +7,34 @@
 </head>
 <body>
 
+  <div class="container-fluid">
 
-  <!-- Modal -->
-  <div class="modal-header">
-    <!-- <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button> -->
-    <h4 class="modal-title" id="myModalLabel">Add URLs</h4>
+    <br />
+
+    <div class="row">
+      <div class="col-xs-9">
+        <textarea placeholder="URLs" id="urlinput"></textarea>
+
+        <p class="small text-center subtle">
+          Many online services are supported.<br />
+          You can add more than one URL.<br />
+          To add a new line press <code>Shift</code> + <code>Enter</code>.
+        </p>
+      </div>
+      <div class="col-xs-3">
+        <p class="text-center">
+          <button type="button" id="urlAddBtn" class="btn btn-md btn-primary" data-toggle="addurl" data-addurl-input="urlinput">Add</button>
+        </p>
+      </div>
+    </div>
+
   </div>
-
-  <div class="modal-body">
-    <!-- <input type="text" value="" placeholder="URL" id="urlinput" /> -->
-    <textarea placeholder="URLs" id="urlinput"></textarea>
-
-    <p class="small text-center subtle">Various online services are supported. You can add multiple entries at once (each on a new line).</p>
-    <p class="small text-center subtle">Tip: you can add with <code>Ctrl/Cmd</code> + <code>Enter</code>.</p>
-  </div>
-
-  <div class="modal-footer">
-    <!-- <button type="button" class="btn btn-sm btn-default" data-dismiss="modal">Cancel</button> -->
-    <button type="button" class="btn btn-sm btn-warning" data-toggle="addurl" data-addurl-input="urlinput">Add</button>
-  </div>
-
 
   <assets style="display: none">
     <link rel="icon" type="image/png" href="icons/player-256.png">
     <link rel="apple-touch-icon" href="icons/player-256.png">
     <link rel="stylesheet" href="vendor/css/bootstrap.min.css">
     <link rel="stylesheet" href="css/base.css">
-    <link rel="stylesheet" href="css/player.css">
     <link rel="stylesheet" href="css/themes.css">
     <script>window.$ = window.jQuery = require('./vendor/js/jquery-2.2.4.min.js');</script>
