+++ views/css/base.css
@@ -1,6 +1,6 @@
 /* Defaults */
 html, body, .wrap, .containerBody, .sidebar, .mainContent {height: 100%}
-body {background: transparent; font-size: 12px}
+body {background: #eee; font-size: 12px}
 * {cursor: default}
 table.table {font-size: 1em}
 
@@ -8,13 +8,10 @@ table.table {font-size: 1em}
 .padding-10 {padding:10px}
 .padding-0-10 {padding:0 10px}
 
-input.urlviewer {width: 100%; background: none; border: none}
-
 .well {background-color: rgba(66, 66, 66, 0.85); border: 1px solid rgba(0, 0, 0, 0.3)}
 .well a {color: white}
 
 .wrap {padding: 0; margin: 0}
-.entry-url {display: block; font-size: 0.8em; font-weight: bold}
 
 /* List group */
 .list-group-item {background-color: transparent; border: none; color: #555}
@@ -68,7 +65,6 @@ ul.list li .primary {color: black; text-decoration: none}
 
 
 
-/* NEW WINDOW CONCEPT */
 body {/*-webkit-app-region: drag;*/ -webkit-user-select: none; cursor: default}
 
 
@@ -126,11 +122,15 @@ label.settings-item select {width:80px}
 
 
 
-/* PREFERENCES */
+/* Position */
 #wCtls.right {position: fixed; left: auto; right: 8px}
 #wCtls.right .windowCtl {float: right}
 
-
+/* Styling */
 .windowCtl i {display: none}
 #wCtls.generic .windowCtl i.generic {display: inline-block;}
 #wCtls.mac .windowCtl i.mac {display: inline-block;}
+
+
+textarea#urlinput {width: 100%; height:150px; resize:none; outline: none}
+#urlAddBtn {width: 100%; height:70px; margin:40px 0}
