+++ views/help.html
@@ -15,6 +15,13 @@
     <div id="updateNotice"></div>
   </div>
 
+  <h5>Feedback? Issues?</h5>
+  <hr />
+
+  <p>If you encounter any problems please <a href="https://github.com/vot/remotestack-player/issues">create an issue on GitHub</a>.</p>
+
+  <br /><br />
+
   <h5>Keyboard shortcuts: Player control</h5>
   <hr />
 
@@ -72,7 +79,6 @@
     <link rel="apple-touch-icon" href="icons/player-256.png">
     <link rel="stylesheet" href="vendor/css/bootstrap.min.css">
     <link rel="stylesheet" href="css/base.css">
-    <link rel="stylesheet" href="css/player.css">
     <link rel="stylesheet" href="css/themes.css">
     <script>window.$ = window.jQuery = require('./vendor/js/jquery-2.2.4.min.js');</script>
