+++ views/settings.html
@@ -11,7 +11,7 @@
   <!-- Theme -->
   <label class="settings-item">
     <span class="pull-right">
-      <select name="ui.theme" onchange="RS.UI.handlePreferencesInputChange(this)">
+      <select name="ui.theme" onchange="RS.SettingsWindow.handlePreferencesInputChange(this)">
         <option></option>
         <option value="light">Light</option>
         <option value="dark">Dark</option>
@@ -23,7 +23,7 @@
   <!-- Window controls position -->
   <label class="settings-item">
     <span class="pull-right">
-      <select name="ui.windowCtlPosition" onchange="RS.UI.handlePreferencesInputChange(this)">
+      <select name="ui.windowCtlPosition" onchange="RS.SettingsWindow.handlePreferencesInputChange(this)">
         <option></option>
         <option value="left">Left</option>
         <option value="right">Right</option>
@@ -35,7 +35,7 @@
   <!-- Window controls style -->
   <label class="settings-item">
     <span class="pull-right">
-      <select name="ui.windowCtlStyle" onchange="RS.UI.handlePreferencesInputChange(this)">
+      <select name="ui.windowCtlStyle" onchange="RS.SettingsWindow.handlePreferencesInputChange(this)">
         <option></option>
         <option value="generic">Generic</option>
         <option value="mac">Mac</option>
@@ -47,7 +47,7 @@
   <!-- Show full path -->
   <label class="settings-item">
     <span class="pull-right">
-      <input type="checkbox" value="ui.showFullPath" onchange="RS.UI.handlePreferencesCheckboxChange(this)" />
+      <input type="checkbox" value="ui.showFullPath" onchange="RS.SettingsWindow.handlePreferencesCheckboxChange(this)" />
     </span>
     <strong>Show full path in playlist</strong>
   </label>
@@ -55,7 +55,7 @@
   <!-- Developer mode -->
   <label class="settings-item">
     <span class="pull-right">
-      <input type="checkbox" value="app.enableDeveloperMode" onchange="RS.UI.handlePreferencesCheckboxChange(this)" />
+      <input type="checkbox" value="app.enableDeveloperMode" onchange="RS.SettingsWindow.handlePreferencesCheckboxChange(this)" />
     </span>
     <strong>Developer mode (requires restart)</strong>
   </label>
@@ -70,7 +70,6 @@
     <link rel="apple-touch-icon" href="icons/player-256.png">
     <link rel="stylesheet" href="vendor/css/bootstrap.min.css">
     <link rel="stylesheet" href="css/base.css">
-    <link rel="stylesheet" href="css/player.css">
     <link rel="stylesheet" href="css/themes.css">
     <script>window.$ = window.jQuery = require('./vendor/js/jquery-2.2.4.min.js');</script>
